重构(cli)：使用clap和模块重构应用结构
整个应用之前仅包含在一个`main.rs`文件中，现已重构以提高可维护性和提供更健壮的用户体验。

主要变更包括：
- 使用`clap` crate替换手动命令行参数解析，提供更好的帮助信息、验证和更符合Rust习惯的CLI。
- 将单一结构分解为专用模块：
  - `cli.rs`：定义命令结构。
  - `handlers.rs`：实现每个命令的逻辑。
  - `utils.rs`：包含共享的辅助函数。
- 使用`colored` crate增强终端输出，提高可读性。
