+++ src/main.rs
@@ -1,458 +1,23 @@
-use std::fs;
-use std::process::Command;
-use std::path::Path;
-use std::collections::HashSet;
-use std::sync::mpsc::channel;
-use notify::{Watcher, RecursiveMode};
+mod cli;
+mod handlers;
+mod utils;
+
+use clap::Parser;
+use cli::{Cli, Commands};
+use handlers::*;
 
 fn main() {
-    let args: Vec<String> = std::env::args().collect();
-    
-    match args.len() {
-        1 => show_menu(),
-        2 => {
-            let command = &args[1];
-            match command.as_str() {
-                "list" => list_exercises(),
-                "watch" => watch_mode(),
-                "hint" => show_general_hint(),
-                "reset" => reset_progress(),
-                "progress" => show_progress(),
-                _ => {
-                    check_exercise(command);
-                }
-            }
-        }
-        3 => {
-            if args[1] == "hint" {
-                show_hint(&args[2]);
-            } else {
-                show_usage();
-            }
-        }
-        _ => show_usage(),
+    let cli = Cli::parse();
+
+    match cli.command {
+        Some(Commands::List) => list_exercises(),
+        Some(Commands::Run { name }) => {
+            check_exercise(&name);
+        },
+        Some(Commands::Hint { name }) => show_hint(&name),
+        Some(Commands::Reset) => reset_progress(),
+        Some(Commands::Progress) => show_progress(),
+        Some(Commands::Watch) => watch_mode(),
+        None => show_menu(),
     }
-}
-
-fn show_menu() {
-    println!("🏃‍♂️ Welcome to Movelings!");
-    println!("=======================");
-    println!();
-    println!("💡 Commands:");
-    println!("  cargo run <exercise>     - Check specific exercise");
-    println!("  cargo run list          - List all exercises");
-    println!("  cargo run progress      - Show detailed progress");
-    println!("  cargo run hint <exercise> - Get hint for exercise");
-    println!("  cargo run reset         - Reset progress");
-    println!("  cargo run watch         - Watch mode (auto-check on save)");
-    println!();
-    
-    let exercises = get_exercises();
-    let completed = load_completed_exercises();
-    
-    println!("📚 Quick overview:");
-    println!("  Total exercises: {}", exercises.len());
-    println!("  Completed: {}", completed.len());
-    println!("  Remaining: {}", exercises.len() - completed.len());
-    
-    if !completed.is_empty() {
-        println!("  Progress: {:.1}%", (completed.len() as f64 / exercises.len() as f64) * 100.0);
-    }
-    
-    println!();
-    println!("🚀 Start with: cargo run 00_intro");
-    println!("📋 To see all exercises: cargo run list");
-}
-
-fn list_exercises() {
-    let exercises = get_exercises();
-    let completed = load_completed_exercises();
-    
-    println!("📚 Available exercises:");
-    
-    for (i, exercise) in exercises.iter().enumerate() {
-        let icon = if completed.contains(exercise) { "✅" } else { "📝" };
-        println!("  {}. {} {} (run 'cargo run {}' to test)", i + 1, icon, exercise, exercise);
-    }
-    
-    println!();
-    println!("📊 Progress: {}/{} completed ({:.1}%)", 
-             completed.len(), 
-             exercises.len(), 
-             (completed.len() as f64 / exercises.len() as f64) * 100.0);
-    println!("💡 Icons: ✅ = Completed, 📝 = Not completed yet");
-    println!("💡 To get hints: cargo run hint <exercise_name>");
-}
-
-fn get_exercises() -> Vec<String> {
-    let mut exercises = Vec::new();
-    if let Ok(entries) = fs::read_dir("exercises") {
-        for entry in entries {
-            if let Ok(entry) = entry {
-                if entry.path().is_dir() {
-                    exercises.push(entry.file_name().to_string_lossy().to_string());
-                }
-            }
-        }
-    }
-    exercises.sort();
-    exercises
-}
-
-fn check_exercise(name: &str) -> bool {
-    let exercise_path = format!("exercises/{}", name);
-    if !Path::new(&exercise_path).exists() {
-        println!("❌ Exercise '{}' not found!", name);
-        println!("Run 'cargo run list' to see available exercises.");
-        return false;
-    }
-    
-    println!("🔍 Checking exercise: {}", name);
-    println!("📁 Path: {}", exercise_path);
-    
-    println!("⏳ Running tests... (this may take a moment)");
-    
-    let output = Command::new("sui")
-        .args(&["move", "test"])
-        .current_dir(&exercise_path)
-        .output();
-    
-    match output {
-        Ok(output) => {
-            if output.status.success() {
-                println!("✅ Exercise '{}' completed successfully!", name);
-                
-                // 保存完成记录
-                save_completed_exercise(name);
-                
-                println!("🎉 Great job! Try the next exercise.");
-                suggest_next_exercise(name);
-                true
-            } else {
-                println!("❌ Exercise '{}' failed:", name);
-                let stderr = String::from_utf8_lossy(&output.stderr);
-                let stdout = String::from_utf8_lossy(&output.stdout);
-                
-                if !stderr.is_empty() {
-                    println!("\n📋 Error output:");
-                    println!("{}", stderr);
-                }
-                if !stdout.is_empty() {
-                    println!("\n📋 Test output:");
-                    println!("{}", stdout);
-                }
-                
-                show_hint(name);
-                false
-            }
-        }
-        Err(e) => {
-            println!("❌ Failed to run sui command: {}", e);
-            println!("💡 Make sure Sui CLI is installed and in your PATH");
-            println!("💡 You can also try running 'sui move test' manually in the exercise directory");
-            false
-        }
-    }
-}
-
-fn suggest_next_exercise(current: &str) {
-    let exercises = get_exercises();
-    let completed = load_completed_exercises();
-    
-    if let Some(pos) = exercises.iter().position(|x| x == current) {
-        if pos + 1 < exercises.len() {
-            println!("➡️  Next exercise: cargo run {}", exercises[pos + 1]);
-        } else {
-            println!("🎊 Congratulations! You've completed all exercises!");
-            println!("🏆 Final progress: {}/{} exercises completed!", completed.len(), exercises.len());
-        }
-    }
-}
-
-fn load_completed_exercises() -> HashSet<String> {
-    let mut completed = HashSet::new();
-    if let Ok(content) = fs::read_to_string(".movelings_progress") {
-        for line in content.lines() {
-            let line = line.trim();
-            if !line.is_empty() {
-                completed.insert(line.to_string());
-            }
-        }
-    }
-    completed
-}
-
-fn save_completed_exercise(exercise: &str) {
-    let mut completed = load_completed_exercises();
-    if completed.insert(exercise.to_string()) {
-        // 只有当这是新完成的练习时才保存
-        save_progress(&completed);
-        println!("💾 Progress saved!");
-    }
-}
-
-fn save_progress(completed: &HashSet<String>) {
-    let mut exercises: Vec<String> = completed.iter().cloned().collect();
-    exercises.sort(); // 保持有序
-    
-    if let Err(e) = fs::write(".movelings_progress", exercises.join("\n")) {
-        eprintln!("⚠️  Warning: Failed to save progress: {}", e);
-    }
-}
-
-fn reset_progress() {
-    if Path::new(".movelings_progress").exists() {
-        match fs::remove_file(".movelings_progress") {
-            Ok(_) => println!("🔄 Progress reset successfully!"),
-            Err(e) => println!("❌ Failed to reset progress: {}", e),
-        }
-    } else {
-        println!("ℹ️  No progress file found, nothing to reset.");
-    }
-}
-
-fn show_progress() {
-    let exercises = get_exercises();
-    let completed = load_completed_exercises();
-    
-    println!("📊 Detailed Progress Report");
-    println!("==========================");
-    println!();
-    
-    if exercises.is_empty() {
-        println!("No exercises found!");
-        return;
-    }
-    
-    println!("📚 Exercise Status:");
-    for (i, exercise) in exercises.iter().enumerate() {
-        let status = if completed.contains(exercise) {
-            "✅ Completed"
-        } else {
-            "📝 Not completed"
-        };
-        println!("  {}. {} - {}", i + 1, exercise, status);
-    }
-    
-    println!();
-    println!("📈 Statistics:");
-    println!("  Total exercises: {}", exercises.len());
-    println!("  Completed: {}", completed.len());
-    println!("  Remaining: {}", exercises.len() - completed.len());
-    println!("  Progress: {:.1}%", (completed.len() as f64 / exercises.len() as f64) * 100.0);
-    
-    if completed.len() == exercises.len() {
-        println!();
-        println!("🎉 Congratulations! You've completed all exercises!");
-    } else if let Some(next) = exercises.iter().find(|ex| !completed.contains(*ex)) {
-        println!();
-        println!("➡️  Next exercise to try: cargo run {}", next);
-    }
-}
-
-fn show_hint(exercise: &str) {
-    println!();
-    println!("💡 Hints for exercise '{}':", exercise);
-    
-    // 尝试从文件中提取提示
-    let hints = extract_hints_from_exercise(exercise);
-    
-    if hints.is_empty() {
-        // 如果没有找到提示，使用默认提示
-        show_default_hints();
-    } else {
-        for (i, hint) in hints.iter().enumerate() {
-            println!("  {}. {}", i + 1, hint);
-        }
-    }
-    
-    println!();
-    println!("📝 Exercise file location:");
-    println!("   exercises/{}/sources/", exercise);
-}
-
-fn extract_hints_from_exercise(exercise: &str) -> Vec<String> {
-    let mut hints = Vec::new();
-    let sources_path = format!("exercises/{}/sources", exercise);
-    
-    if let Ok(entries) = fs::read_dir(&sources_path) {
-        for entry in entries {
-            if let Ok(entry) = entry {
-                if let Some(ext) = entry.path().extension() {
-                    if ext == "move" {
-                        if let Ok(content) = fs::read_to_string(entry.path()) {
-                            hints.extend(extract_hints_from_content(&content));
-                        }
-                    }
-                }
-            }
-        }
-    }
-    
-    hints
-}
-
-fn extract_hints_from_content(content: &str) -> Vec<String> {
-    let mut hints = Vec::new();
-    let mut seen = std::collections::HashSet::new();
-    
-    for line in content.lines() {
-        let line = line.trim();
-        let mut hint_text = None;
-        
-        // 提取以 "HINT:" 开头的注释
-        if line.starts_with("// HINT:") || line.starts_with("/// HINT:") {
-            let hint = line.trim_start_matches("// HINT:")
-                          .trim_start_matches("/// HINT:")
-                          .trim();
-            if !hint.is_empty() {
-                hint_text = Some(hint.to_string());
-            }
-        }
-        
-        // 提取 TODO 项目
-        else if line.starts_with("// TODO:") || line.starts_with("/// TODO:") {
-            let todo = line.trim_start_matches("// TODO:")
-                          .trim_start_matches("/// TODO:")
-                          .trim();
-            if !todo.is_empty() {
-                hint_text = Some(format!("TODO: {}", todo));
-            }
-        }
-        
-        // 提取编号的提示 (// 1. xxx, // 2. xxx)
-        else if line.starts_with("//") {
-            let comment = line.trim_start_matches("//").trim();
-            if comment.len() > 2 && 
-               comment.chars().nth(0).unwrap_or(' ').is_ascii_digit() && 
-               comment.chars().nth(1) == Some('.') {
-                hint_text = Some(comment.to_string());
-            }
-        }
-        
-        // 避免重复
-        if let Some(hint) = hint_text {
-            if seen.insert(hint.clone()) {
-                hints.push(hint);
-            }
-        }
-    }
-    
-    hints
-}
-
-fn show_default_hints() {
-    let hints = vec![
-        "Read the exercise file comments carefully",
-        "Look for TODO markers that guide you",
-        "Check the test functions to understand what's expected",
-        "Try running 'sui move test' manually to see detailed errors",
-    ];
-    
-    for (i, hint) in hints.iter().enumerate() {
-        println!("  {}. {}", i + 1, hint);
-    }
-}
-
-fn show_general_hint() {
-    println!("💡 General hints:");
-    println!("  - Read exercise comments carefully");
-    println!("  - Look for TODO markers");
-    println!("  - Check test functions for expected behavior");
-    println!("  - Use 'cargo run hint <exercise>' for specific help");
-}
-
-fn watch_mode() {
-    println!("👀 Entering watch mode. Press Ctrl+C to exit.");
-    println!("📂 Watching for changes in the 'exercises' directory...");
-
-    let (tx, rx) = channel();
-
-    let mut watcher = match notify::recommended_watcher(move |res| {
-        if let Ok(event) = res {
-            if let Err(e) = tx.send(event) {
-                println!("❌ Error sending event: {}", e);
-            }
-        } else if let Err(e) = res {
-            println!("❌ Watch error: {}", e);
-        }
-    }) {
-        Ok(w) => w,
-        Err(e) => {
-            println!("❌ Failed to create watcher: {}", e);
-            return;
-        }
-    };
-
-    if let Err(e) = watcher.watch(Path::new("exercises"), RecursiveMode::Recursive) {
-        println!("❌ Failed to start watching 'exercises' directory: {}", e);
-        return;
-    }
-
-    // Initial check of the first uncompleted exercise
-    let exercises = get_exercises();
-    let completed = load_completed_exercises();
-    if let Some(first_exercise) = exercises.iter().find(|e| !completed.contains(*e)) {
-        println!("\n🚀 Starting with the first uncompleted exercise: {}", first_exercise);
-        check_exercise(first_exercise);
-        println!("----------------------------------------------------");
-        println!("👀 Watching for changes... Press Ctrl+C to exit.");
-    }
-
-
-    loop {
-        match rx.recv() {
-            Ok(event) => {
-                if let notify::Event {
-                    kind: notify::EventKind::Modify(_),
-                    paths,
-                    ..
-                } = event
-                {
-                    for path in paths {
-                        let is_in_sources = path.components().any(|c| c.as_os_str() == "sources");
-                        let is_in_build = path.components().any(|c| c.as_os_str() == "build");
-                        let is_move_file = path.extension().map_or(false, |s| s == "move");
-
-                        if is_move_file && is_in_sources && !is_in_build {
-                            if let Some(exercise_name) = find_exercise_name_from_path(&path) {
-                                // Clear screen for better readability
-                                print!("\x1B[2J\x1B[1;1H");
-                                println!("----------------------------------------------------");
-                                println!("✨ Detected change in: {}", path.display());
-                                check_exercise(&exercise_name);
-                                println!("----------------------------------------------------");
-                                println!("👀 Watching for next change... Press Ctrl+C to exit.");
-                            }
-                        }
-                    }
-                }
-            }
-            Err(e) => {
-                println!("❌ Error receiving event: {}", e);
-                break;
-            }
-        }
-    }
-}
-
-fn find_exercise_name_from_path(path: &Path) -> Option<String> {
-    path.ancestors()
-        .find(|p| {
-            p.parent().map_or(false, |parent| parent.file_name() == Some("exercises".as_ref()))
-        })
-        .and_then(|p| p.file_name())
-        .and_then(|s| s.to_str())
-        .map(|s| s.to_string())
-}
-
-fn show_usage() {
-    println!("Usage: cargo run [command] [exercise]");
-    println!("Commands:");
-    println!("  cargo run                    - Show main menu");
-    println!("  cargo run <exercise>         - Test specific exercise");
-    println!("  cargo run list              - List all exercises with status");
-    println!("  cargo run progress          - Show detailed progress");
-    println!("  cargo run hint <exercise>   - Get hints for exercise");
-    println!("  cargo run reset             - Reset all progress");
 }
